Refactor to use core namespace
